Fixed: App cannot start if no Network is availible (Update Checker BUg)

+++ SFDL.NET 3/Modules/UpdateCheck.vb
@@ -64,7 +64,9 @@ Module UpdateCheck
         Catch ex As Exception
             _log.Error(ex, ex.Message)
         Finally
-            IO.File.Delete(_tmp_file)
+            If IO.File.Exists(_tmp_file) Then
+                IO.File.Delete(_tmp_file)
+            End If
         End Try
 
         Return _rt
